# rev_core/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Reverse Proxy Manager - LLM Gateway",
    "description": "Secure reverse proxy for OpenAI-compatible upstream backends. Routes requests by model, enforces token-level quotas, and records usage.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "Send token via Authorization: Bearer <TOKEN>"
      }
    },
...
  },
  "paths": {
...
    "/v1/models": {
      "get": {
        "summary": "List models",
        "description": "Forwards to upstream /v1/models via the first active provider.",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Model list", "content": { "application/json": { "schema": { "type": "object" } } } },
          "401": { "description": "Missing or invalid token" }
        }
      }
    },
    "/v1/models/{model_id}": {
      "get": {
        "summary": "Get model",
        "description": "Forwards to upstream /v1/models/{model_id} via the matching provider.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Model metadata", "content": { "application/json": { "schema": { "type": "object" } } } },
          "401": { "description": "Missing or invalid token" }
        }
      }
    },
...
  }
}